Fix some printf format warnings.

+++ daemon-manager.cc
@@ -384,12 +384,12 @@ static string do_command(string command_line, user *user, vector<class daemon*> 
         string resp = strprintf("%-30s %-15s %6s %8s %8s %6s %6s\n", "daemon-id", "state", "pid", "respawns", "cooldown", "uptime", "total");
         for (vector<class daemon*>::iterator d = manageable.begin(); d != manageable.end(); d++)
           if (arg.empty() || arg == (*d)->id)
-            resp += strprintf("%-30s %-15s %6d %8d %8d %6d %6d\n",
+            resp += strprintf("%-30s %-15s %6d %8zd %8d %6ld %6ld\n",
                               (*d)->id.c_str(),
                               (*d)->state_str().c_str(),
                               (*d)->pid,
                               (*d)->respawns,
-                              (*d)->cooldown_remaining(),
+                              (int)(*d)->cooldown_remaining(),
                               (*d)->pid ? time(NULL) - (*d)->respawn_time : 0,
                               (*d)->pid ? time(NULL) - (*d)->start_time   : 0);
         return "OK: " + resp;
